공유 버튼 클릭 시 Firebase 딥링크로 앱/스토어 분기 이동 : refactor : 주석 수정, url 인코딩처리, app-config 위치 변경  https://github.com/TEAM-ROMROM/RomRom-FE/issues/765
